feat: update contribution contract 'new_contribution' call signature refactor: domain type GithubIssueNumber is now u64 instead of i64

+++ marketplace-core/src/routes/contributions/create.rs
@@ -17,7 +17,7 @@ pub async fn create_contribution(
 
 	let contribution = ContributionProjection {
 		project_id: body.project_id(),
-		issue_number: body.github_issue_number(),
+		issue_number: body.github_issue_number() as u64,
 		gate: body.gate(),
 		..Default::default()
 	};
